tests: Rename `./go test server`, move tests
Makes the structure and scope of the tests and their accompanying
scripts easier to follow. Also streamlined `scripts/test` slightly.
