+++ scripts/test.d/end-to-end
@@ -30,7 +30,7 @@ _test_end_to_end() {
       "available Node.js version is $node_version. Skipping..."$'\n' >&2
     return
   fi
-  mocha tests/end-to-end.js
+  mocha tests/end-to-end/end-to-end.js
 }
 
 _test_end_to_end "$@"
